requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
discord.py==2.5.2
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2.1
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
gunicorn==23.0.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.1
whitenoise==6.9.0
yarl==1.20.0
